feat: Ask for a valid input if entered one is invalid
If the entered input in the prompt is invalid then a message appears to alert about it & then the user is required to enter a vailid one. If the input is again invalid, the previously mentioned steps are repeated until a valid input is entered or the process is canceled.